refactor: simplify chain configuration
