+++ src/components/ChainSwitcher/index.tsx
@@ -1,29 +1,26 @@
 import { Box, Button } from '@mui/material'
+import { hexValue } from 'ethers/lib/utils'
 import type { ReactElement } from 'react'
 
-import { useChains } from '@/hooks/useChains'
+import { useChain } from '@/hooks/useChain'
 import { useOnboard } from '@/hooks/useOnboard'
 import { useChainId } from '@/hooks/useChainId'
 import { useWallet } from '@/hooks/useWallet'
-import { switchWalletChain } from '@/utils/wallet'
 
 import css from './styles.module.css'
 
 export const ChainSwitcher = (): ReactElement | null => {
   const onboard = useOnboard()
   const wallet = useWallet()
-  const { data: chains } = useChains()
+  const defaultChain = useChain()
   const chainId = useChainId()
-  const defaultChain = chains?.results.find((chain) => chain.chainId === chainId.toString())
 
   if (!wallet || wallet.chainId === chainId.toString()) {
     return null
   }
 
   const handleChainSwitch = async () => {
-    if (onboard) {
-      await switchWalletChain(onboard, wallet, chainId)
-    }
+    onboard?.setChain({ chainId: hexValue(chainId) })
   }
 
   return (
