+++ src/hooks/useEnsResolution.ts
@@ -3,10 +3,9 @@ import { getAddress, isAddress } from 'ethers/lib/utils'
 import { useEffect, useState } from 'react'
 
 import { parsePrefixedAddress, sameAddress } from '@/utils/addresses'
-import { useChains } from '@/hooks/useChains'
+import { useChain } from '@/hooks/useChain'
 import { useWeb3 } from '@/hooks/useWeb3'
 import { useWallet } from '@/hooks/useWallet'
-import { useChainId } from '@/hooks/useChainId'
 
 export const useEnsResolution = (str: string, debounce = true): [string | undefined, string | undefined, boolean] => {
   const [result, setResult] = useState<string>()
@@ -16,10 +15,7 @@ export const useEnsResolution = (str: string, debounce = true): [string | undefi
   const web3 = useWeb3()
   const wallet = useWallet()
   const { safe } = useSafeAppsSDK()
-  const { data: chains } = useChains()
-  const chainId = useChainId()
-
-  const chain = chains?.results.find((chain) => chain.chainId === chainId.toString())
+  const chain = useChain()
   const shortName = chain?.shortName
 
   useEffect(() => {
